Basic support for Scala 3 (no derivation)
