+++ shared/src/main/scala-2/toml/util/RecordToMap.scala
