+++ shared/src/main/scala-3/toml/TomlVersionSpecific.scala
@@ -1,0 +1,5 @@
+package toml
+
+trait TomlVersionSpecific {
+  
+}
