+++ shared/src/main/scala-3/toml/shapeless.scala
@@ -1,0 +1,11 @@
+package toml
+
+type HList = Tuple
+
+type Lazy[T] = T
+
+// object Witness {
+//   type Aux[T]
+// }
+
+// given Aux[T] = null
